+++ lib/authentication/authentication_wrapper.dart
@@ -1,6 +1,6 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
-import 'package:skinalert/home.dart';
+import 'package:skinalert/home/home.dart';
 import 'package:skinalert/login_or_signup.dart';
